[SECURITY] Protect arguments of form __referrer with HMAC
The request arguments of the referring request are
a serialized string written to one of the hidden
fields in a Fluid form. This string has to be protected
by a HMAC to protect Extbase from possible unserialize
attacks.

Note: For now there is no object known within Extbase,
that could be used for an unserialize exploit!

This change also backports some convenience hmac methods
to the hash service from the current master, to have the
bugfix in sync.

Change-Id: Iaba9e7556c7aeeb8a3724c84d96160a511500e98
Security-Bulletin: TYPO3-CORE-SA-2012-001
Fixes: #35310
Releases: 1.4, 4.7, 6.0