[BUGFIX] Container calls Cache too often
The Container Classes Cache call is very expensive - and when
using a database backend, this is even slowing down
compared to a NullBackend.

One reason is that for every cache request the cache is
called 3 times: 1 time has and 2 times get.
This can be avoided and be reduced to one get call -
which reduces request time by 10-20% for a bigger Extbase request.

Change-Id: I43a0ebf7e0bfd067f3e48c0e7c27f12660b4cb7c
Resolves: #36008
Releases: 4.7, 6.0
Reviewed-on: http://review.typo3.org/12975
Reviewed-by: Tymoteusz Motylewski
Tested-by: Tymoteusz Motylewski
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn