[BUGFIX] Property mapper: Exception not thrown but skipped
The transformToObject function in the property mapper
checks for an existing record if an identity is given.
If no record could be found an exception should be thrown.
But the exception will never be thrown due to a wrong
condition. The condition checks for FALSE but NULL is
given.

The Exception does also not exist.

Change-Id: If6ac4d3275718cabab9c94904dad264472630c18
Fixes: #39876
Releases: 1.3, 1.4, 4.7, 6.0
Reviewed-on: http://review.typo3.org/16512
Reviewed-by: Ingo Pfennigstorf
Tested-by: Ingo Pfennigstorf